requirements.txt
tqdm
opencv-python
fvcore
torchgeometry
importlib-metadata
chumpy